Changed lib organization for multiple grammars, no newline needed anymore & lint fix
